+++ components/footer.tsx
@@ -1,0 +1,10 @@
+
+const Footer = () => {
+    return (
+        <footer className="flex justify-center items-center h-12 bg-red-900 text-white">
+            &copy; 2021 Delsco Enterprises. All Rights Reserved.
+        </footer>
+    )
+}
+
+export default Footer
